requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.4.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.3
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==1.0.0
Django==3.2.5
django-cors-headers==4.2.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.0.12
fqdn==1.5.1
frozenlist==1.4.0
h11==0.14.0
idna==2.10
importlib-metadata==6.0.0
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
lxml==4.6.3
MarkupSafe==2.1.2
mistune==2.0.5
multidict==6.0.4
mysqlclient==2.1.1
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
outcome==1.2.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2021.5.29
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
psutil==5.9.4
psycopg2==2.9.6
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.8.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2021.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.25.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
selenium==4.10.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.2.1
sqlparse==0.4.1
stack-data==0.6.2
starlette==0.27.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
twilio==8.5.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.6
uvicorn==0.22.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
whitenoise==6.2.0
widgetsnbextension==4.0.5
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
